整理代码

+++ data/views.py
@@ -1,18 +1,20 @@
 from django.http import JsonResponse
-from django.shortcuts import render
+from requests import Response
+from rest_framework.decorators import api_view
+from rest_framework.views import APIView
 
-# Create your views here.
+from data.serializers import DataSerializer
 from tools.logging_dec import logging_check
 from data.models import Data
 import json
-from user.models import User
-import time
+from django.utils.decorators import method_decorator
 
 
-@logging_check
-def data(request):
-    if request.method == 'GET':
-        data_all = Data.objects.all()
+class PoemThreeAPI(APIView):
+    @method_decorator(logging_check)
+    def get(self, request):
+
+        data_all = Data.objects.all().filter(user_id=request.myuser)
         result = []
         for install in data_all:
             table = {'id': install.id, 'user_id': install.user_id.username, 'url_name': install.url_name,
@@ -23,7 +25,8 @@ def data(request):
         result = {'code': 200, 'result': result}
         return JsonResponse(result)
 
-    elif request.method == 'POST':
+    @method_decorator(logging_check)
+    def post(self, request):
         json_str = request.body
         json_obj = json.loads(json_str)
         url_name = json_obj['url_name']
@@ -39,15 +42,21 @@ def data(request):
         result = {'code': 200}
         return JsonResponse(result)
 
-    elif request.method == 'DELETE':
+    @method_decorator(logging_check)
+    def delete(self, request):
         json_str = request.body
         json_obj = json.loads(json_str)
         data_id = json_obj['data_id']
-        Data.objects.filter(id=data_id).delete()
-        result = {'code': 200}
-        return JsonResponse(result)
+        try:
+            Data.objects.filter(id=data_id, user_id=request.myuser).delete()
+            result = {'code': 200}
+            return JsonResponse(result)
+        except Exception as e:
+            result = {'code': 404}
+            return JsonResponse(result)
 
-    elif request.method == 'PUT':
+    @method_decorator(logging_check)
+    def put(self, request):
         json_str = request.body
         json_obj = json.loads(json_str)
         id = json_obj['data_id']
@@ -56,7 +65,12 @@ def data(request):
         user_name = json_obj['user_name']
         user_password = json_obj['user_password']
         remarks = json_obj['remarks']
-        Data.objects.filter(id=id).update(url_name=url_name, url=url, user_name=user_name, user_password=user_password,
-                                          remarks=remarks)
-        result = {'code': 200}
-        return JsonResponse(result)
+        try:
+            Data.objects.filter(id=id, user_id=request.myuser).update(url_name=url_name, url=url, user_name=user_name,
+                                                                      user_password=user_password,
+                                                                      remarks=remarks)
+            result = {'code': 200}
+            return JsonResponse(result)
+        except Exception as e:
+            result = {'code': 404}
+            return JsonResponse(result)
